Pull BucketLifecycleSettings into its own file.
