+++ src/workspaces/SettingsModal/SettingsModal.tsx
@@ -1,19 +1,10 @@
-import {
-  ButtonPrimary,
-  CreatableSelect,
-  ExternalLink,
-  Modal,
-  SpinnerOverlay,
-  Switch,
-  useUniqueId,
-} from '@terra-ui-packages/components';
+import { ButtonPrimary, Modal, SpinnerOverlay } from '@terra-ui-packages/components';
 import _ from 'lodash/fp';
 import React, { ReactNode, useEffect, useState } from 'react';
-import { NumberInput } from 'src/components/input';
 import { Ajax } from 'src/libs/ajax';
 import { withErrorReporting } from 'src/libs/error';
-import { FormLabel } from 'src/libs/forms';
 import { useCancellation } from 'src/libs/react-utils';
+import BucketLifecycleSettings from 'src/workspaces/SettingsModal/BucketLifecycleSettings';
 import {
   BucketLifecycleSetting,
   DeleteBucketLifecycleRule,
@@ -21,15 +12,10 @@ import {
   isDeleteBucketLifecycleRule,
   modifyFirstBucketDeletionRule,
   removeFirstBucketDeletionRule,
+  suggestedPrefixes,
   WorkspaceSetting,
 } from 'src/workspaces/SettingsModal/utils';
 import { WorkspaceWrapper as Workspace } from 'src/workspaces/utils';
-
-export const suggestedPrefixes = {
-  allObjects: 'All Objects',
-  submissions: 'submissions/',
-  submissionIntermediaries: 'submissions/intermediates/',
-};
 
 interface SettingsModalProps {
   workspace: Workspace;
@@ -48,9 +34,6 @@ const SettingsModal = (props: SettingsModalProps): ReactNode => {
 
   // Original settings from server, may contain multiple types
   const [workspaceSettings, setWorkspaceSettings] = useState<WorkspaceSetting[] | undefined>(undefined);
-
-  const switchId = useUniqueId('switch');
-  const daysId = useUniqueId('days');
 
   const signal = useCancellation();
 
@@ -101,14 +84,8 @@ const SettingsModal = (props: SettingsModalProps): ReactNode => {
     return !!deleteRules && deleteRules.length >= 1 ? deleteRules[0] : undefined;
   };
 
-  const prefixOptions = () => {
-    // Append together suggested options any options the user has added or are already selected.
-    const allOptions = _.uniq(_.concat(_.values(suggestedPrefixes), prefixes));
-    return _.map((value) => ({ value, label: value }), allOptions);
-  };
-
   const persistSettings = async () => {
-    let newSettings;
+    let newSettings: WorkspaceSetting[];
     if (lifecycleRulesEnabled) {
       const prefixesOrNone = _.without([suggestedPrefixes.allObjects], prefixes);
       newSettings = modifyFirstBucketDeletionRule(workspaceSettings || [], lifecycleAge!, prefixesOrNone);
@@ -135,83 +112,14 @@ const SettingsModal = (props: SettingsModalProps): ReactNode => {
         </ButtonPrimary>
       }
     >
-      <div style={{ display: 'flex', flexDirection: 'row', alignItems: 'center', marginTop: '6px' }}>
-        <FormLabel
-          htmlFor={switchId}
-          style={{ fontWeight: 'bold', whiteSpace: 'nowrap', marginRight: '5px', marginTop: '5px' }}
-        >
-          Lifecycle Rules:
-        </FormLabel>
-        <Switch
-          onLabel=''
-          offLabel=''
-          onChange={(checked: boolean) => {
-            setLifecycleRulesEnabled(checked);
-            if (!checked) {
-              // Clear out the values being display to reduce
-              setLifecycleAge(null);
-              setPrefixes([]);
-            }
-          }}
-          id={switchId}
-          checked={lifecycleRulesEnabled}
-          width={40}
-          height={20}
-        />
-      </div>
-
-      <div style={{ marginTop: '20px', marginBottom: '10px' }}>
-        <div style={{ marginTop: '5px', marginBottom: '5px' }}>Delete objects in:</div>
-        <CreatableSelect
-          isClearable
-          isMulti
-          isSearchable
-          placeholder='Choose "All Objects" or specify prefixes'
-          aria-label='Specify if all objects should be deleted, or objects with specific prefixes'
-          value={_.map((value) => ({ value, label: value }), prefixes)}
-          onChange={(data) => {
-            const selectedPrefixes = _.map('value', data);
-            // If added "All Objects", clear the others.
-            if (
-              _.contains(suggestedPrefixes.allObjects, selectedPrefixes) &&
-              !_.contains(suggestedPrefixes.allObjects, prefixes)
-            ) {
-              setPrefixes([suggestedPrefixes.allObjects]);
-            } else if (selectedPrefixes.length > 1) {
-              setPrefixes(_.without([suggestedPrefixes.allObjects], selectedPrefixes));
-            } else {
-              setPrefixes(selectedPrefixes);
-            }
-          }}
-          options={prefixOptions()}
-          isDisabled={!lifecycleRulesEnabled}
-        />
-      </div>
-      <div style={{ display: 'flex', alignItems: 'center' }}>
-        {/* eslint-disable jsx-a11y/label-has-associated-control */}
-        <label style={{ marginRight: '5px' }} htmlFor={daysId}>
-          Days after upload:
-        </label>
-        <NumberInput
-          style={{ maxWidth: '100px' }}
-          id={daysId}
-          min={0}
-          isClearable
-          onlyInteger
-          value={lifecycleAge}
-          disabled={!lifecycleRulesEnabled}
-          onChange={(value: number) => {
-            setLifecycleAge(value);
-          }}
-        />
-      </div>
-      {/* TODO: add aria-described by to link to the setting. */}
-      <div style={{ marginTop: '10px', fontSize: '12px' }}>
-        This{' '}
-        <ExternalLink href='https://cloud.google.com/storage/docs/lifecycle'>bucket lifecycle setting</ExternalLink>{' '}
-        automatically deletes objects a certain number of days after they are uploaded.{' '}
-        <span style={{ fontWeight: 'bold' }}>Changes can take up to 24 hours to take effect.</span>
-      </div>
+      <BucketLifecycleSettings
+        lifecycleRulesEnabled={lifecycleRulesEnabled}
+        setLifecycleRulesEnabled={setLifecycleRulesEnabled}
+        lifecycleAge={lifecycleAge}
+        setLifecycleAge={setLifecycleAge}
+        prefixes={prefixes}
+        setPrefixes={setPrefixes}
+      />
       {workspaceSettings === undefined && <SpinnerOverlay />}
     </Modal>
   );
